+++ src/main/java/com/crp/ucp/character/CharacterRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<CharacterEntity, Long> {
+public interface CharacterRepository extends JpaRepository<CharacterEntity, Integer> {
 
     Optional<CharacterEntity> findByName(String name);
 }
